# docs/prompts/api.yaml
openapi: 3.0.0
info:
    title: Chat-API
    version: 1.0.0
    description: API endpoint for prompts in the Chat-API platform.
    contact:
        email: hello@aimpact.partners

paths:
    /prompts/templates/{id}:
        get:
            summary: Get a Prompt.
            tags:
                - Prompts Templates
            responses:
                200:
                    description: Prompt Geted successfully.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponse'
                400:
                    description: Bad request. Possible errors regarding missing or incorrect data.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponseBadRequest'

        put:
            summary: Update a Prompt.
            tags:
                - Prompts Templates
            responses:
                200:
                    description: Prompt Geted successfully.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponse'
                400:
                    description: Bad request. Possible errors regarding missing or incorrect data.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponseBadRequest'

        delete:
            summary: Delete a prompt.
            tags:
                - Prompts Templates
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - value
                            properties:
                                value:
                                    type: string

            responses:
                200:
                    description: Prompt published successfully.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponse'
                400:
                    description: Bad request. Possible errors regarding missing or incorrect data.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponseBadRequest'

    /prompts/templates:
        post:
            summary: Publish a new prompt.
            tags:
                - Prompts Templates
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            allOf:
                                - $ref: '#/components/schemas/IPromptTemplateDataParams'
                                - type: object
                            required:
                                - name
                                - description
                                - language
                            properties:
                                categories:
                                    type: array
                                    items:
                                        type: string

            responses:
                200:
                    description: Prompt published successfully.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponse'
                400:
                    description: Bad request. Possible errors regarding missing or incorrect data.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponseBadRequest'

    /prompts/templates/project/{id}:
        get:
            summary: Get a prompts templates by project.
            tags:
                - Prompts Templates
            parameters:
                - name: is
                  in: query
                  description: Filter ('prompt', 'function', 'dependency').
                  schema:
                      type: string
                      enum:
                          - prompt
                          - function
                          - dependency
            responses:
                200:
                    description: Prompt Geted successfully.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponseList'
                400:
                    description: Bad request. Possible errors regarding missing or incorrect data.
                    content:
                        application/json:
                            schema:
                                type: object
                                $ref: '#/components/schemas/IResponseBadRequest'

components:
    schemas:
        IResponse:
            type: object
            required:
                - status
            properties:
                status:
                    type: boolean
                    description: Response status.
                data:
                    type: object
                    $ref: '#/components/schemas/IPromptTemplateData'
                    description: Response Data.

        IResponseBadRequest:
            type: object
            properties:
                status:
                    type: boolean
                    description: Response status.
                error:
                    type: string
                    description: Error text.

        IResponseList:
            type: object
            required:
                - status
            properties:
                status:
                    type: boolean
                    description: Response status.
                data:
                    type: object
                    properties:
                        entries:
                            description: Response Data.
                            type: array
                            items:
                                $ref: '#/components/schemas/IPromptTemplateData'

        IPromptTemplateDataParams:
            allOf:
                - $ref: '#/components/schemas/IPromptTemplateBaseData'
                - type: object
            type: object
            required:
                - projectId
                - name
                - description
                - language
                - format
                - is
            properties:
                projectId:
                    type: string

        IPromptTemplateBaseData:
            allOf:
                - $ref: '#/components/schemas/IPromptTemplateData'
                - type: object
            type: object
            properties:
                options:
                    type: array
                    items:
                        $ref: '#/components/schemas/IPromptTemplateOptionData'

        IPromptTemplateData:
            type: object
            required:
                - id
                - name
                - description
                - language
                - format
                - is
            properties:
                id:
                    type: string
                name:
                    type: string
                identifier:
                    type: string
                description:
                    type: string
                language:
                    type: string
                format:
                    type: string
                    enum:
                        - json
                        - text
                is:
                    type: string
                    enum:
                        - prompt
                        - function
                        - dependency
                value:
                    type: string
                options:
                    type: array
                    items:
                        $ref: '#/components/schemas/IPromptTemplateOptionData'
                literals:
                    type: object
                    properties:
                        pure:
                            type: array
                            items:
                                type: string
                        dependencies:
                            type: array
                            items:
                                type: string
                        metadata:
                            type: array
                            items:
                                type: string

        IPromptTemplateCategoryData:
            properties:
                id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                prompts:
                    type: object
                    additionalProperties:
                        type: string

        IPromptTemplateOptionData:
            properties:
                id:
                    type: string
                value:
                    type: string